Restructure and renaming of hash classed
Add folder for hash module
Rename ChainHashTable -> ChainedHashTable
Fix tests
