+++ lib/data_structures_101.rb
@@ -1,6 +1,8 @@
 require "data_structures_101/version"
 require "data_structures_101/linked_list"
-require "data_structures_101/hash_table"
+require "data_structures_101/hash/base_hash_table"
+require "data_structures_101/hash/bucket"
+require "data_structures_101/chained_hash_table"
 
 module DataStructures101
   # Your code goes here...
